Split developer and PM MyTickets page
1. Fixed getTicketsByDeveloperEmail method;
2. Change ConditionalRender to function instead of React component
3. Split MyTickets in PM-Tickets and Developer-Tickets
